refactor: Rename and move shared folder to lib/src/
