+++ lib/main.dart
@@ -5,7 +5,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:namaadhu_vaguthu/services/data_service.dart';
 import 'package:namaadhu_vaguthu/providers/global_providers.dart';
 import 'package:namaadhu_vaguthu/services/notification_service.dart';
-import 'package:namaadhu_vaguthu/shared/constants.dart';
+import 'package:namaadhu_vaguthu/src/constants/constants.dart';
 import 'package:namaadhu_vaguthu/src/utils/prayertimes_utils.dart';
 import 'package:namaadhu_vaguthu/src/utils/string_utils.dart';
 import 'package:namaadhu_vaguthu/src/utils/time_utils.dart';
