+++ lib/screens/custom_search_delegate.dart
@@ -4,7 +4,7 @@ import 'package:ionicons/ionicons.dart';
 import 'package:namaadhu_vaguthu/models/atoll.dart';
 import 'package:namaadhu_vaguthu/models/island.dart';
 import 'package:namaadhu_vaguthu/providers/selected_island_provider.dart';
-import 'package:namaadhu_vaguthu/shared/constants.dart';
+import 'package:namaadhu_vaguthu/src/constants/constants.dart';
 
 class CustomSearchDelegate extends SearchDelegate {
   final List<Island> islandList;
